requirements.txt
asgiref==3.5.1
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.14
django-cors-headers==3.5.0
djangorestframework==3.13.1
djongo==1.3.6
dnspython==2.2.1
drf-yasg==1.20.0
greenlet==1.1.2
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
mongo==0.2.0
packaging==21.3
pymongo==3.12.1
pyparsing==3.0.9
pytz==2022.1
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
SQLAlchemy==1.4.36
sqlparse==0.2.4
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
